Refactored backup functionality to be more scalable.
